Temp fix for ImAlive to fall back to MG22 when MG21 is not returning "ok"
This update provides a temporary work-around for MG21 unduly returning "ko" for active feeders. This "ko" message is because the MongoDB database on MG21, which maintains the feeder status, isn't correctly updated. 
Temp solution is for every "ko" to then also poll MG22, which generally will provide the correct status. 

This results in a double poll of MG21 and MG22 in case MG21 returns "ko", which is acceptable for now because these polls happen only in 10 minute intervals for active stations.

Once the MongoDB issue is fixed on MG21, this code becomes harmless because the extra poll will only be done if a feeder is actually "bad". This would be only a very small subsets of all stations feeding to MG21. In any case, we advise to revert this change once the ImAlive function on MG21 is fixed and stable.